requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
click==8.0.3
cycler==0.10.0
falcon==3.0.1
h11==0.12.0
importlib-metadata==4.8.1
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
opencv-python==4.5.4.58
Pillow==8.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
uvicorn==0.15.0
zipp==3.6.0
zxing==0.13.1